example/requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
hermes @ git+git://github.com/PaymeTrade/hermes.git@97561aca4cd22714d2796280c0054edefd7a9677
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.1
mccabe==0.6.1
numpy==1.19.2
pandas==1.1.3
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.11
websocket-client==0.56.0
wrapt==1.12.1
xmpppy==0.6.1
